[SP] Leave CGhoul2Info_v::size returning int for now because a lot of code depends on it.

+++ code/game/ghoul2_shared.h
@@ -459,7 +459,7 @@ public:
 		}
 		Array().push_back(model);
 	}
-	size_t size() const
+	int size() const
 	{
 		if (!IsValid())
 		{
